--- a/SharpPcap/WinDivert/WinDivertNative.cs
+++ b/SharpPcap/WinDivert/WinDivertNative.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 using System.Runtime.InteropServices;
 
 namespace SharpPcap.WinDivert
@@ -12,7 +11,16 @@ namespace SharpPcap.WinDivert
         internal static extern IntPtr WinDivertOpen([MarshalAs(UnmanagedType.LPStr)] string filter, WinDivertLayer layer, short priority, ulong flags);
 
         [DllImport(WINDIVERT_DLL, CallingConvention = CallingConvention.Cdecl, SetLastError = true)]
-        internal static extern bool WinDivertRecv(IntPtr handle, IntPtr pPacket, uint packetLen, out uint readLen, out WinDivertAddress pAddr);
+        internal static extern bool WinDivertRecvEx(
+            IntPtr handle,
+            [In, Out] byte[] pPacket,
+            int packetLen,
+            out int pRecvLen,
+            ulong flags,
+            [In, Out] WinDivertAddress[] pAddr,
+            ref int pAddrLen,
+            IntPtr lpOverlapped
+        );
 
         [DllImport(WINDIVERT_DLL, CallingConvention = CallingConvention.Cdecl, SetLastError = true)]
         internal static extern bool WinDivertClose(IntPtr handle);