Pending reply on https://github.com/basil00/Divert/issues/282 to finish implementation 

Partially resolves #273