@kayoub5 2021-09-04
@trudyhood I believe you have an application/setup to test WinDivert performance since you reported #273, could you check this PR?
I had to add a call to `WinDivertHelperParsePacket` so I am not 100% confident on the performance improvement