@trudyhood 2021-09-04
@kayoub5
Thank you for the improvement  you have done but unfortunately the result is disappointing,
The new SharpPcap speed, is still close to "passthru.exe true 1 1", it looks there is no noticeable improvement,  as you see "passthru.exe true 1 10" cab boost its own performance more than 2X.
here is the results:

### No Filter
```
   Speedtest by Ookla

     Server: NETGEAR Inc. - San Jose, CA (id = 43447)
        ISP: Amazon.com
    Latency:     0.25 ms   (0.03 ms jitter)
   Download:   963.65 Mbps (data used: 449.4 MB)
     Upload:   608.07 Mbps (data used: 717.7 MB)
Packet Loss: Not available.
 Result URL: https://www.speedtest.net/result/c/779c63f1-1830-482a-a35a-3dafe55fe27b
```

### Old SharpPcap
```
   Speedtest by Ookla

     Server: NETGEAR Inc. - San Jose, CA (id = 43447)
        ISP: Amazon.com
    Latency:     0.37 ms   (0.02 ms jitter)
   Download:   140.88 Mbps (data used: 257.7 MB)
     Upload:   163.18 Mbps (data used: 279.5 MB)
Packet Loss: Not available.
 Result URL: https://www.speedtest.net/result/c/0d5117ad-cda1-496d-ad2f-8a4598ffa634
```

### New SharpPcap with batch
```
   Speedtest by Ookla

     Server: NETGEAR Inc. - San Jose, CA (id = 43447)
        ISP: Amazon.com
    Latency:     0.53 ms   (0.02 ms jitter)
   Download:   160.86 Mbps (data used: 277.0 MB)
     Upload:   158.79 Mbps (data used: 213.9 MB)
https://www.speedtest.net/result/c/cbf5253e-6b61-4981-9d6b-de0959379f99
```

###  Original WinDivert Batch 1 : passthru.exe true 1 1
```
     Speedtest by Ookla

     Server: NETGEAR Inc. - San Jose, CA (id = 43447)
        ISP: Amazon.com
    Latency:     0.38 ms   (0.03 ms jitter)
   Download:   217.66 Mbps (data used: 340.4 MB)
     Upload:   203.27 Mbps (data used: 272.8 MB)
Packet Loss: Not available.
 Result URL: https://www.speedtest.net/result/c/6e8c7dfd-6778-4235-8ebe-45a6cf39a0df
```

###  Original WinDivert  Batch 10 : passthru.exe true 1 10
Speedtest by Ookla

```
     Server: NETGEAR Inc. - San Jose, CA (id = 43447)
        ISP: Amazon.com
    Latency:     0.32 ms   (0.03 ms jitter)
   Download:   490.18 Mbps (data used: 866.2 MB)
     Upload:   335.49 Mbps (data used: 580.5 MB)
Packet Loss: Not available.
 Result URL: https://www.speedtest.net/result/c/d3176b31-b8bb-4c58-90fb-17bfe5fd3c35
```
I used the following code:

```c#
using PacketDotNet;
using SharpPcap.WinDivert;
using System;
using System.Threading;

namespace SimpleFilter
{
    class Program
    {
        static readonly WinDivertDevice WinDivert = new();
        private static long _length;
        private static long _lastLength;

        static void Main()
        {

            WinDivert.Layer = WinDivertLayer.Network;
            WinDivert.Filter = "ip";
            WinDivert.Flags = 0;
            WinDivert.Open(new SharpPcap.DeviceConfiguration());
            WinDivert.OnPacketArrival += WinDivert_OnPacketArrival;

            Console.WriteLine("Working...");
            WinDivert.StartCapture();
            Thread.Sleep(-1);
        }

        private static void WinDivert_OnPacketArrival(object sender, SharpPcap.PacketCapture e)
        {
            var rawPacket = e.GetPacket();
            var packet = Packet.ParsePacket(rawPacket.LinkLayerType, rawPacket.Data);
            var lastCaptureHeader = (WinDivertHeader)e.Header;
            var ipPacket = packet.Extract<IPv4Packet>();

            WinDivert.SendPacket(ipPacket.Bytes, lastCaptureHeader);

            _length += packet.TotalPacketLength;
            if (_length > _lastLength + 1000000)
            {
                _lastLength = _length;
                Console.WriteLine(_lastLength);
            }
        }
    }
}

```