@trudyhood 2021-09-04
@kayoub5
I am not sure the issue is rooted in .NET overhead, but as we see, SharpPcap performance is close to "passthru.exe true 1 1."
I looked at your code; you capture packets in batch, but to be compatible with SharpPacp, you needed to send them one by one, so eventually, we couldn't pass the packets in batch using WinDivertSendEx. As I used tcp connection for speed so decreasing send speed lead to decrease in receive speed too.

I wish SharpPcap.PacketCapture contains an array of packets instead of a single packet but I think it would be big change!
Can you test and create a .NET passthru using WinDivertRecvEx, WinDivertSendEx. so we can check the feasibility!