@trudyhood 2021-09-04
> @trudyhood I only implemented WinDivertRecvEx, sending is still being done packet per packet
> that could be the bottleneck in your case

Ya, I saw your message after sending my own.