@kayoub5 2021-09-04
@trudyhood could you profile your application and upload the results, my ISP is not fast enough to test 200Mbps 😞 

See https://docs.microsoft.com/en-us/visualstudio/profiling/cpu-usage?view=vs-2019

Instrumentation, CPU usage and .NET Object Allocation Tracking  are the interesting reports in this case