@trudyhood 2021-09-04
> @trudyhood could you profile your application and upload the results, my ISP is not fast enough to test 200Mbps 😞
> The report would help identify the bottleneck, so we don't have to speculate
> 
> See https://docs.microsoft.com/en-us/visualstudio/profiling/cpu-usage?view=vs-2019
> 
> Instrumentation, CPU usage and .NET Object Allocation Tracking are the interesting reports in this case

@kayoub5 My ISP is not that fast; I compiled it on my machine and ran it on an Amazon EC2 server with 1Gbit speed. Unfortunately, it just has 1GB of memory which is not enough to install visual studio.