@kayoub5 2021-09-04
Just checked with profiler with 40Mbps traffic, most of the CPU time goes to either `WinDivertRecvEx` or `WinDivertSend`, `WinDivertHelperParsePacket` is fast and consume almost no CPU time.

So the good news is that this PR introduce no regression performance, and should allow faster reception of packets.

The not so good news is that we need some kind of API for `WinDivertSendEx`