@trudyhood 2021-09-06
> This improves performance when capturing using windivert? I saw the messages on the ticket but didn't read them in depth to confirm this was the case.

I believe it is much more to be done, the improvement is not noticeable yet.